+++ react/components/ShippingSimulator/components/ShippingTableRow.js
@@ -4,12 +4,14 @@ import { intlShape, injectIntl } from 'react-intl'
 import TranslateEstimate from 'vtex.shipping-estimate-translator/TranslateEstimate'
 import classNames from 'classnames'
 
+import styles from '../styles'
+
 const ShippingTableRow = ({ name, shippingEstimate, price, intl }) => {
-  const etaClassName = classNames('vtex-shipping-table__cell pv1 ph3 t-small c-muted-2', {
+  const etaClassName = classNames(`${styles.shippingTableCell} pv1 ph3 t-small c-muted-2`, {
     'tc': shippingEstimate === undefined,
   })
 
-  const valueClassName = classNames('vtex-shipping-table__cell pv1 ph3 t-small c-muted-2', {
+  const valueClassName = classNames(`${styles.shippingTableCell} pv1 ph3 t-small c-muted-2`, {
     'tc': price === undefined,
   })
 
@@ -20,15 +22,15 @@ const ShippingTableRow = ({ name, shippingEstimate, price, intl }) => {
   } else if (price === 0) {
     valueText = intl.formatMessage({ id: 'shipping.free' })
   } else {
-    valueText = intl.formatNumber(price/100, { style: 'currency', currency: 'BRL' })
+    valueText = intl.formatNumber(price / 100, { style: 'currency', currency: 'BRL' })
   }
 
   return (
     <tr key={name}>
-      <td className="vtex-shipping-table__cell pv1 ph3 t-small">
-        <label className="vtex-shipping-table__shipping-name-label">
+      <td className={`${styles.shippingTableCell} pv1 ph3 t-small`}>
+        <label className={`${styles.shippingTableLabel}`}>
           <input
-            className="vtex-shipping-table__radio-input mr4"
+            className={`${styles.shippingTableRadioBtn} mr4`}
             name="shipping-option"
             type="radio"
             value={name}