+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support to CSS modules in `ProductDescription`.
 - Support to CSS modules in `Greeting`.
 - Support to CSS modules in `CollectionBadges`.
+- Support to CSS modules in `DiscountBadge`.
 
 ## [3.5.8] - 2019-01-15
 ### Fixed