+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support to CSS modules in `Greeting`.
 - Support to CSS modules in `CollectionBadges`.
 - Support to CSS modules in `DiscountBadge`.
+- Support to CSS modules in `AvailabilitySubscriber`.
 
 ## [3.5.8] - 2019-01-15
 ### Fixed